Refactor of the app, update to new backend spec
 - refactor: cleaner structure and model use
 - couple of tweaks to pass new backend spec tests
 - update of readme as per backend spec
