+++ README.md
@@ -1,2 +1,40 @@
-# example-backend-node-express
-An example backend shunter application using the Express web framework for node.js
+
+# Example Shunter Application Back End (Node.js, Express)
+
+A Node.js-based back end for the [Shunter example application](https://github.com/shunterjs/example).
+
+
+## Requirements
+
+[Node.js](https://nodejs.org/) 0.12+ is required to run this application.
+
+
+## Installation
+
+Install dependencies with `npm install`.
+
+
+## Running
+
+Start the application on port `5000`:
+
+```
+npm start
+```
+
+Start the application on a different port:
+
+```
+PORT=1234 npm start
+```
+
+Start the application with [Foreman](https://github.com/ddollar/foreman):
+
+```
+foreman start
+```
+
+
+## License
+
+Copyright 2016, licensed under the [Lesser General Public License (LGPL-3.0)](http://www.gnu.org/licenses/lgpl-3.0.txt).
