+++ package.json
@@ -24,6 +24,7 @@
     "jshint": "^2"
   },
   "scripts": {
+    "start": "node app.js",
     "lint": "make lint"
   }
 }
